requirements.txt
langchain_community
sqlalchemy
langchain_openai
langchain_community
psycopg2
requests
langchain_groq